TKSS-961: Split kona_sm2.c
