+++ kona-crypto/src/main/jni/kona_sm2.h
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2024, THL A29 Limited, a Tencent company. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
+EVP_PKEY* load_pub_key(const uint8_t* pub_key, size_t pub_key_len);
+
+EVP_PKEY* load_key_pair(const uint8_t* pri_key, const uint8_t* pub_key);
+
+int sm2_gen_pub_key(const uint8_t* pri_key, uint8_t* pub_key);
+
+EVP_PKEY_CTX* sm2_create_pkey_ctx(EVP_PKEY* pkey);
